move TeamRelatedViewset to own file; social_view stuff inherits from base class
